requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
pydantic==1.11.1
requests==2.28.1
python-dotenv==0.21.1
loguru==0.6.0
